# swagger.yaml
openapi: 3.0.3
info:
  title: Jigi pay
  description: >-
    # A simple payment app api
  version: 1.0.0
  contact: { email: ammuftau74@gmail.com, whatapp: +2340935095173 }
servers:
  - url: https://jigipay.onrender.com/api/v1
    description: This is the original api URL
    schemes: [https, http]
  - url: http://localhost:8080/api/v1
    description: THis is my local Server
paths:
...
  /client/cards:
    post:
      tags:
        - 📝 Client
      summary: Add card
      description: >-
        The Add Card API allows logged-in users to add back card to their account.
      operationId: addCard
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  example: 897jii8u97089e
                name:
                  type: string
                  example: james king
                number:
                  type: string
                  example: 1374455465655
                expe:
                  type: string
                  example: 12/24
                expire:
                  type: number
                  example: 124
            example:
              userId: 13744524r5344
              name: monday david
              number: 13743432435
              expire: 12/34
              cvv: 134
      responses:
        '200':
          description: Add account card
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '85'
            Date:
              schema:
                type: string
                example: Wed, 31 May 2023 19:59:47 GMT
            ETag:
              schema:
                type: string
                example: W/"55-CyGJCaxqcCXrzMloar9z7m95mvk"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties: {}
                  message:
                    type: string
                    example: Card added successfully
                  success:
                    type: boolean
                    example: true
              examples:
                Add account card:
                  value:
                    data: {}
                    message: Card added successfully
                    success: true
    get:
      tags:
        - 📝 Client
      summary: Get card
      description: >-
        The API endpoint retrieves a list of card.
      operationId: getCards
      responses:
        '200':
          description: Get cards
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '10757'
            Date:
              schema:
                type: string
                example: Thu, 01 Jun 2023 05:44:13 GMT
            ETag:
              schema:
                type: string
                example: W/"2a05-xjkqHU4sNAeFPZX9uGNel7XPn8E"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            _id:
                              type: string
                              example: 649e795264e7dba29b726936
                            userId:
                              type: string
                              example: 649e795264e7dba29b726936
                            number:
                              type: number
                              example: 89974059435
                            cvv:
                              type: string
                              example: 649e795264e7dba29b726336
                            name:
                              type: string
                              example: david muda
                            expire:
                              type: string
                              example: 649e795264e7dba29b726336
                        example:
                          _id: 68poi09ii99iou8946359
                          userId: 68poi09ii99iou8946359
                          cvv: 689
                          expire: 12/20
                          number: 08909789789
                          name: Joseph
                  message:
                    type: string
                    example: Card fetched successfully
                  success:
                    type: boolean
                    example: true
              examples:
                # edit title
                Get card:
                  value:
                    data:
                      data:
                        - userId: 649e795264e7dba29b96936
                          _id: 649e795264e7dba29b726036
                          name: jame monday
                          number: 987948749
                          cvv: 7897
                          expire: 12/26
                        - _id: 649e795264e7dba29b726036
                          userId: 649e795264e7dba29b96936
                          name: musa kano
                          number: 123120
                          cvv: 274
                          expire: 03/25
                    message: Card fetched successfully
                    success: true
    delete:
      tags:
        - 📝 Client
      summary: Delete user cards
      description: >-
        The API endpoint allows you to delete cards.
      operationId: deleteCards
      responses:
        '200':
          description: Delete user card by id
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '443'
            Date:
              schema:
                type: string
                example: Fri, 30 Jun 2023 06:44:47 GMT
            ETag:
              schema:
                type: string
                example: W/"1bb-YVaffq7t8o3of0yDl9g26tprr/8"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            RateLimit-Limit:
              schema:
                type: string
                example: '500'
            RateLimit-Remaining:
              schema:
                type: string
                example: '493'
            RateLimit-Reset:
              schema:
                type: string
                example: '631'
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: cards deleted succefully
                      status:
                        type: string
                        example: success
...
  /client/budgets:
    post:
      tags:
        - 📝 Client
      summary: Add budget
      description: >-
        The Add budget API allows logged-in users to add back budget for their account.
      operationId: addBudget
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Valentine
                amount:
                  type: string
                  example: 137
                startDate:
                  type: date
                  example: 12/24
                expire:
                  endDate: date
                  example: 124
                type:
                  type: fun
                  example: 137
                cycle:
                  type: fun
                  example: 137
                frequency:
                  type: fun
                  example: 137
            example:
              name: valentine
              amount: 200
              startDate: 1/12/2024
              endDate: 12/12/2024
              type: fun
              frequency: 12/12/2024
              cycle: 12/12/2024
      responses:
        '200':
          description: Add budget
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '85'
            Date:
              schema:
                type: string
                example: Wed, 31 May 2023 19:59:47 GMT
            ETag:
              schema:
                type: string
                example: W/"55-CyGJCaxqcCXrzMloar9z7m95mvk"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties: {}
                  message:
                    type: string
                    example: budget added successfully
                  success:
                    type: boolean
                    example: true
              examples:
                Add budget:
                  value:
                    data: {}
                    message: budget added successfully
                    success: true
    get:
      tags:
        - 📝 Client
      summary: Get budget
      description: >-
        The API endpoint retrieves a list of budget.
      operationId: getBudgets
      responses:
        '200':
          description: Get budgets
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '10757'
            Date:
              schema:
                type: string
                example: Thu, 01 Jun 2023 05:44:13 GMT
            ETag:
              schema:
                type: string
                example: W/"2a05-xjkqHU4sNAeFPZX9uGNel7XPn8E"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            _id:
                              type: string
                              example: 649e795264e7dba29b726936
                            userId:
                              type: string
                              example: 649e795264e7dba29b726936
                            number:
                              type: number
                              example: 89974059435
                            cvv:
                              type: string
                              example: 649e795264e7dba29b726336
                            name:
                              type: string
                              example: david muda
                            expire:
                              type: string
                              example: 649e795264e7dba29b726336
                        example:
                          _id: 68poi09ii99iou8946359
                          userId: 68poi09ii99iou8946359
                          cvv: 689
                          expire: 12/20
                          number: 08909789789
                          name: Joseph
                  message:
                    type: string
                    example: budget fetched successfully
                  success:
                    type: boolean
                    example: true
              examples:
                # edit title
                Get card:
                  value:
                    data:
                      data:
                        - userId: 649e795264e7dba29b96936
                          _id: 649e795264e7dba29b726036
                          name: jame monday
                          number: 987948749
                          cvv: 7897
                          expire: 12/26
                        - _id: 649e795264e7dba29b726036
                          userId: 649e795264e7dba29b96936
                          name: musa kano
                          number: 123120
                          cvv: 274
                          expire: 03/25
                    message: Card fetched successfully
                    success: true
...